+++ src/rer/ufficiostampa/events/comunicato.py
@@ -14,12 +14,19 @@ def setNumber(item, event):
     setattr(item, "comunicato_number", get_next_comunicato_number())
 
 
-def setEmptyNumber(item, event):
+def createComunicato(item, event):
     """
-    Reset it when copy a comunicato
+    Reset it when copy a comunicato and force set legislature
     """
     setattr(item, "comunicato_number", "")
     setattr(item, "message_sent", False)
+
+    # this is needed because it's a readonly field and that doesn't store anything
+    # in the content. Side effect is that it also have a defaultFactory value that
+    # always return the latest legislature. When you try to edit an item for a past
+    # legislature, you always get the latest one, instead the "stored" one.
+
+    setattr(item, "legislature", getattr(item, "legislature", ""))
 
 
 def fixText(item, event):